+++ backend/src/main/java/com/celuveat/auth/presentation/AuthConstant.java
@@ -1,4 +1,4 @@
-package com.celuveat.common.auth;
+package com.celuveat.auth.presentation;
 
 public class AuthConstant {
